Renkodat efter marcus lektion
