+++ HenriksHobbyLager/HenriksHobbyLager.csproj
@@ -8,10 +8,4 @@
     <RootNamespace>HenriksHobbyLager</RootNamespace>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Compile Include="Database\dummy.md" />
-    <Compile Include="Interfaces\dummy.md" />
-    <Compile Include="Models\dummy.md" />
-  </ItemGroup>
-
 </Project>
